requirements.txt
crewai>=0.1.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
GitPython>=3.1.0
pydantic>=2.0.0
pytest>=7.0.0
litellm>=1.0.0

# requirements.txt
# Run: pip install -r requirements.txt

# Web framework + server
fastapi
uvicorn[standard]       # [standard] includes websockets and http tools

